update 'tunnel' plugin for latest cfoundry APIs
